.Queried for completeness flags from Dynamics based on an ApplicationLine rather than a Ruleset Added Waste Recovery Plan page (although this isn't included in the current permit we are working on).
