+++ src/views/taskList.html
@@ -22,7 +22,7 @@
 
         <ol class="task-list">
 
-          {{#each standardRule.sections}}
+          {{#each taskList.sections}}
           <li>
             <h2 class="task-list-section">
               <span id="{{this.id}}-number" class="task-list-section-number">{{this.sectionNumber}}. </span><span id="{{this.id}}-heading">{{this.sectionName}}</span>
