+++ src/views/wasteRecoveryPlan.html
@@ -1,0 +1,9 @@
+<main id="content" role="main">
+
+  {{> common/betaBanner }}
+
+  {{> common/backLink }}
+
+  <h1 id="waste-recovery-plan-heading" class="heading-large">{{ pageHeading }}</h1>
+
+</main>
